requirements.txt
panns-inference
torchaudio
tqdm
transformers